move ek80 unit tests to test_calibrate_ek80
